src/main/java/BackEndC3/ClinicaOdontologica/Service/OdontologoService.java
package BackEndC3.ClinicaOdontologica.Service;

import BackEndC3.ClinicaOdontologica.Entity.Odontologo;
import BackEndC3.ClinicaOdontologica.Repository.OdontologoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class OdontologoService {
    @Autowired
    private OdontologoRepository odontologoRepository;

    public Odontologo registrarOdontologo(Odontologo odontologo){
        return odontologoRepository.save(odontologo);

    }
    public Optional<Odontologo> buscarPorId(Long id){
        return odontologoRepository.findById(id);
    }
    public Optional<Odontologo> buscarPorMatricula(String matricula){
        return odontologoRepository.findByNumeroMatricula(matricula);
    }
    public List<Odontologo> listarTodos(){
        return odontologoRepository.findAll();
    }
    public void actualizarOdontologo(Odontologo odontologo){
        odontologoRepository.save(odontologo);
    }
    public void eliminarOdontologo(Long id){
        odontologoRepository.deleteById(id);
    }

}
